+++ src/main/java/duke/Parser.java
@@ -1,3 +1,14 @@
+package duke;
+
+import duke.command.Command;
+import duke.command.DeadlineCommand;
+import duke.command.DeleteCommand;
+import duke.command.DoneCommand;
+import duke.command.EventCommand;
+import duke.command.ExitCommand;
+import duke.command.ListCommand;
+import duke.command.TodoCommand;
+
 public class Parser {
     public static Command parse(String input) {
         String inputCommand = input.split(" ")[0];
