+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
+import duke.task.Task;
+
 public class DeleteCommand extends Command {
     public DeleteCommand(String[] args) {
         super(args);
