+++ src/main/java/duke/command/TodoCommand.java
@@ -1,10 +1,16 @@
-public class EventCommand extends Command {
-    public EventCommand(String[] args) {
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
+public class TodoCommand extends Command {
+    public TodoCommand(String[] args) {
         super(args);
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.addEventTask(args[0], args[1]);
+        tasks.addTodoTask(args[0]);
         storage.writeToFile(tasks);
         ui.showNewTask(tasks.lastTask(), tasks.size());
     }
